+++ tests/services/engine/frame/passes/earthLayer.test.ts
@@ -52,10 +52,12 @@ import type { BodyState } from '../../../../../src/@types/scene/BodyState';
 import type { EarthSurfaceTileDrawArgs } from '../../../../../src/@types/rendering/EarthSurfaceTileRenderer';
 
 // Mock composeBodyMvp so the test can (a) assert which vp it consumed by
-// object identity and (b) hand the renderer a recognisable Float32Array. The
-// real composition math is covered by composeBodyMvp's own tests.
+// object identity and (b) hand the layer a recognisable Float64Array — the
+// real (unmocked) composeBodyMvp returns f64; the layer narrows its own copy
+// at the GPU-upload boundary before packEarthSurfaceUniforms. The real
+// composition math is covered by composeBodyMvp's own tests.
 vi.mock('../../../../../src/utils/camera/composeBodyMvp', () => ({
-  composeBodyMvp: vi.fn<() => Float32Array>(() => new Float32Array(16)),
+  composeBodyMvp: vi.fn<() => Float64Array>(() => new Float64Array(16)),
 }));
 import { composeBodyMvp } from '../../../../../src/utils/camera/composeBodyMvp';
